+++ bot_modules/__init__.py
@@ -1,4 +1,9 @@
 # version 1.3.0
 from .snapshots import create_snapshot, list_snapshots, delete_snapshot, restore_snapshot, download_snapshot
+from .verify import verify
+from .utils import get_public_ip, ping
+from .info import info, info_snapshots
 
-__all__ = ['create_snapshot', 'list_snapshots', 'delete_snapshot', 'restore_snapshot', 'download_snapshot']
+__all__ = ['create_snapshot', 'list_snapshots', 'delete_snapshot',
+           'restore_snapshot', 'download_snapshot', 'verify', 'get_public_ip',
+           'info', 'info_snapshots', 'ping']
